Nonsegregating LMNA family data

A dataset containing idiopathic DCM pedigrees with nonsegregating LMNA variants as well as other potentially relevant variants along with quantitative endophenotype measurements. Note that there were no twins in this data set.

lmna_nonseg

Format

A data.table with 88 rows and 13 columns:

family_ID: Family ID

proband: 1 = Yes, 0 = No

individual_ID: Individual identifier (unique within family)

paternal_ID: Individual_ID of person's father

maternal_ID: Individual_ID of person's mother

sex: 1 = Male, 2 = Female

n_lmna_vars: # potentially relevant LMNA variants

n_oth_vars: # potentially relevant non-LMNA variants

grade: Phenotypic severity grade, 0 - 4 (see Cowan et al. 2018 reference)

age_echo_yrs: Age at diagnostic echo, in years

lvedd_z: Left ventricular end-diastolic dimension z-score (standardized for height and sex)

lvef: Left ventricular ejection fraction, \

mzpair: Same value for individuals who are MZ twins; NA otherwise

dzpair: Same value for individuals who are DZ twins; NA otherwise

Source

https://bit.ly/lmna_nonseg

References

Cowan JR, Kinnamon DD, Morales A, Salyer L, Nickerson DA, Hershberger RE. Multigenic Disease and Bilineal Inheritance in Dilated Cardiomyopathy Is Illustrated in Nonsegregating LMNA Pedigrees. Circ Genom Precis Med. 2018 Jul;11(7):e002038. https://doi.org/10.1161/CIRCGEN.117.002038